+++ hugo-nginx/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix documentation for missing parameters
 * Add ports to consul config
 * Adjust hugo-pre step to create new hugo site if not existing
+* Fix git fatal error by moving step to add safe directory before git init step
 
 ---